requirements.txt
asgiref==3.7.2
autopep8==1.4.4
Django==4.2.2
django-ckeditor==6.5.1
django-cors-headers==4.1.0
django-js-asset==2.0.0
django-rest-auth==0.9.5
djangorestframework==3.14.0
dotenv==0.0.5
pep8==1.7.1
Pillow==9.5.0
psycopg2==2.9.6
pycodestyle==2.5.0
python-dotenv==1.0.0
pytz==2018.5
six==1.16.0
sqlparse==0.4.4
tzdata==2023.3
